kalman_filter: fix bug

+++ ros/src/computing/planning/motion/packages/kalman_filter/include/kalman_filter/kalman_filter.h
@@ -5,66 +5,66 @@ class KalmanFilter
 {
 public:
   KalmanFilter();
-  KalmanFilter(const Eigen::MatrixXf &x, const Eigen::MatrixXf &A,
-               const Eigen::MatrixXf &B, const Eigen::MatrixXf &C,
-               const Eigen::MatrixXf &Q, const Eigen::MatrixXf &R,
-               const Eigen::MatrixXf &P);
+  KalmanFilter(const Eigen::MatrixXd &x, const Eigen::MatrixXd &A,
+               const Eigen::MatrixXd &B, const Eigen::MatrixXd &C,
+               const Eigen::MatrixXd &Q, const Eigen::MatrixXd &R,
+               const Eigen::MatrixXd &P);
   ~KalmanFilter();
 
-  void init(const Eigen::MatrixXf &x, const Eigen::MatrixXf &A,
-            const Eigen::MatrixXf &B, const Eigen::MatrixXf &C,
-            const Eigen::MatrixXf &Q, const Eigen::MatrixXf &R,
-            const Eigen::MatrixXf &P);
-  void init(const Eigen::MatrixXf &x, const Eigen::MatrixXf &P0);
-  void setA(const Eigen::MatrixXf &A);
-  void setB(const Eigen::MatrixXf &B);
-  void setC(const Eigen::MatrixXf &C);
-  void setQ(const Eigen::MatrixXf &Q);
-  void setR(const Eigen::MatrixXf &R);
-  void getX(Eigen::MatrixXf &x);
-  void getP(Eigen::MatrixXf &P);
+  void init(const Eigen::MatrixXd &x, const Eigen::MatrixXd &A,
+            const Eigen::MatrixXd &B, const Eigen::MatrixXd &C,
+            const Eigen::MatrixXd &Q, const Eigen::MatrixXd &R,
+            const Eigen::MatrixXd &P);
+  void init(const Eigen::MatrixXd &x, const Eigen::MatrixXd &P0);
+  void setA(const Eigen::MatrixXd &A);
+  void setB(const Eigen::MatrixXd &B);
+  void setC(const Eigen::MatrixXd &C);
+  void setQ(const Eigen::MatrixXd &Q);
+  void setR(const Eigen::MatrixXd &R);
+  void getX(Eigen::MatrixXd &x);
+  void getP(Eigen::MatrixXd &P);
   double getXelement(unsigned int i);
 
-  void predictXandP(const Eigen::MatrixXf &x, const Eigen::MatrixXf &P);
-  void predictEKF(const Eigen::MatrixXf &x_next, const Eigen::MatrixXf &A,
-                  const Eigen::MatrixXf &Q);
-  void predictEKF(const Eigen::MatrixXf &x_next, const Eigen::MatrixXf &A);
-  void predict(const Eigen::MatrixXf &u, const Eigen::MatrixXf &A,
-               const Eigen::MatrixXf &B, const Eigen::MatrixXf &Q);
-  void predict(const Eigen::MatrixXf &u);
+  void predictXandP(const Eigen::MatrixXd &x, const Eigen::MatrixXd &P);
+  void predictEKF(const Eigen::MatrixXd &x_next, const Eigen::MatrixXd &A,
+                  const Eigen::MatrixXd &Q);
+  void predictEKF(const Eigen::MatrixXd &x_next, const Eigen::MatrixXd &A);
+  void predict(const Eigen::MatrixXd &u, const Eigen::MatrixXd &A,
+               const Eigen::MatrixXd &B, const Eigen::MatrixXd &Q);
+  void predict(const Eigen::MatrixXd &u);
 
-  void updateEKF(const Eigen::MatrixXf &y, const Eigen::MatrixXf &y_pred,
-                 const Eigen::MatrixXf &C, const Eigen::MatrixXf &R);
-  void updateEKF(const Eigen::MatrixXf &y, const Eigen::MatrixXf &y_pred,
-                 const Eigen::MatrixXf &C);
-  void update(const Eigen::MatrixXf &y, const Eigen::MatrixXf &C,
-              const Eigen::MatrixXf &R);
-  void update(const Eigen::MatrixXf &y);
+  void updateEKF(const Eigen::MatrixXd &y, const Eigen::MatrixXd &y_pred,
+                 const Eigen::MatrixXd &C, const Eigen::MatrixXd &R);
+  void updateEKF(const Eigen::MatrixXd &y, const Eigen::MatrixXd &y_pred,
+                 const Eigen::MatrixXd &C);
+  void update(const Eigen::MatrixXd &y, const Eigen::MatrixXd &C,
+              const Eigen::MatrixXd &R);
+  void update(const Eigen::MatrixXd &y);
 
 
 private:
-  Eigen::MatrixXf x_;
-  Eigen::MatrixXf A_;
-  Eigen::MatrixXf B_;
-  Eigen::MatrixXf C_;
-  Eigen::MatrixXf Q_;
-  Eigen::MatrixXf R_;
-  Eigen::MatrixXf P_;
+  Eigen::MatrixXd x_;
+  Eigen::MatrixXd A_;
+  Eigen::MatrixXd B_;
+  Eigen::MatrixXd C_;
+  Eigen::MatrixXd Q_;
+  Eigen::MatrixXd R_;
+  Eigen::MatrixXd P_;
 };
 
 KalmanFilter::KalmanFilter() {}
-KalmanFilter::KalmanFilter(const Eigen::MatrixXf &x, const Eigen::MatrixXf &A,
-                           const Eigen::MatrixXf &B, const Eigen::MatrixXf &C,
-                           const Eigen::MatrixXf &Q, const Eigen::MatrixXf &R,
-                           const Eigen::MatrixXf &P)
+KalmanFilter::KalmanFilter(const Eigen::MatrixXd &x, const Eigen::MatrixXd &A,
+                           const Eigen::MatrixXd &B, const Eigen::MatrixXd &C,
+                           const Eigen::MatrixXd &Q, const Eigen::MatrixXd &R,
+                           const Eigen::MatrixXd &P)
 {
   init(x, A, B, C, Q, R, P);
 }
 KalmanFilter::~KalmanFilter() {}
-void KalmanFilter::init(const Eigen::MatrixXf &x, const Eigen::MatrixXf &A,
-                        const Eigen::MatrixXf &B, const Eigen::MatrixXf &C,
-                        const Eigen::MatrixXf &Q, const Eigen::MatrixXf &R,
-                        const Eigen::MatrixXf &P)
+void KalmanFilter::init(const Eigen::MatrixXd &x, const Eigen::MatrixXd &A,
+                        const Eigen::MatrixXd &B, const Eigen::MatrixXd &C,
+                        const Eigen::MatrixXd &Q, const Eigen::MatrixXd &R,
+                        const Eigen::MatrixXd &P)
 {
   x_ = x;
   A_ = A;
@@ -74,68 +74,68 @@ void KalmanFilter::init(const Eigen::MatrixXf &x, const Eigen::MatrixXf &A,
   R_ = R;
   P_ = P;
 }
-void KalmanFilter::init(const Eigen::MatrixXf &x, const Eigen::MatrixXf &P0)
+void KalmanFilter::init(const Eigen::MatrixXd &x, const Eigen::MatrixXd &P0)
 {
   x_ = x;
   P_ = P0;
 }
 
-void KalmanFilter::setA(const Eigen::MatrixXf &A) { A_ = A; }
-void KalmanFilter::setB(const Eigen::MatrixXf &B) { B_ = B; }
-void KalmanFilter::setC(const Eigen::MatrixXf &C) { C_ = C; }
-void KalmanFilter::setQ(const Eigen::MatrixXf &Q) { Q_ = Q; }
-void KalmanFilter::setR(const Eigen::MatrixXf &R) { R_ = R; }
-void KalmanFilter::getX(Eigen::MatrixXf &x) { x = x_; };
-void KalmanFilter::getP(Eigen::MatrixXf &P) { P = P_; };
+void KalmanFilter::setA(const Eigen::MatrixXd &A) { A_ = A; }
+void KalmanFilter::setB(const Eigen::MatrixXd &B) { B_ = B; }
+void KalmanFilter::setC(const Eigen::MatrixXd &C) { C_ = C; }
+void KalmanFilter::setQ(const Eigen::MatrixXd &Q) { Q_ = Q; }
+void KalmanFilter::setR(const Eigen::MatrixXd &R) { R_ = R; }
+void KalmanFilter::getX(Eigen::MatrixXd &x) { x = x_; };
+void KalmanFilter::getP(Eigen::MatrixXd &P) { P = P_; };
 double KalmanFilter::getXelement(unsigned int i) { return x_(i); };
 
-void KalmanFilter::predictXandP(const Eigen::MatrixXf &x, const Eigen::MatrixXf &P)
+void KalmanFilter::predictXandP(const Eigen::MatrixXd &x, const Eigen::MatrixXd &P)
 {
   x_ = x;
   P_ = P;
 }
-void KalmanFilter::predictEKF(const Eigen::MatrixXf &x_next,
-                              const Eigen::MatrixXf &A,
-                              const Eigen::MatrixXf &Q)
+void KalmanFilter::predictEKF(const Eigen::MatrixXd &x_next,
+                              const Eigen::MatrixXd &A,
+                              const Eigen::MatrixXd &Q)
 {
   x_ = x_next;
   P_ = A * P_ * A.transpose() + Q;
 }
-void KalmanFilter::predictEKF(const Eigen::MatrixXf &x_next,
-                              const Eigen::MatrixXf &A)
+void KalmanFilter::predictEKF(const Eigen::MatrixXd &x_next,
+                              const Eigen::MatrixXd &A)
 {
   predictEKF(x_next, A, Q_);
 }
 
-void KalmanFilter::predict(const Eigen::MatrixXf &u, const Eigen::MatrixXf &A,
-                           const Eigen::MatrixXf &B, const Eigen::MatrixXf &Q)
+void KalmanFilter::predict(const Eigen::MatrixXd &u, const Eigen::MatrixXd &A,
+                           const Eigen::MatrixXd &B, const Eigen::MatrixXd &Q)
 {
-  const Eigen::MatrixXf x_next = A * x_ + B * u;
+  const Eigen::MatrixXd x_next = A * x_ + B * u;
   predictEKF(x_next, A, Q);
 }
-void KalmanFilter::predict(const Eigen::MatrixXf &u) { predict(u, A_, B_, Q_); }
+void KalmanFilter::predict(const Eigen::MatrixXd &u) { predict(u, A_, B_, Q_); }
 
-void KalmanFilter::updateEKF(const Eigen::MatrixXf &y,
-                             const Eigen::MatrixXf &y_pred,
-                             const Eigen::MatrixXf &C,
-                             const Eigen::MatrixXf &R)
+void KalmanFilter::updateEKF(const Eigen::MatrixXd &y,
+                             const Eigen::MatrixXd &y_pred,
+                             const Eigen::MatrixXd &C,
+                             const Eigen::MatrixXd &R)
 {
-  const Eigen::MatrixXf S_inv = (R + C * P_ * C.transpose()).inverse();
-  const Eigen::MatrixXf K = P_ * C.transpose() * S_inv;
+  const Eigen::MatrixXd S_inv = (R + C * P_ * C.transpose()).inverse();
+  const Eigen::MatrixXd K = P_ * C.transpose() * S_inv;
   x_ = x_ + K * (y - y_pred);
   const int dim = P_.cols();
   P_ = P_ - (P_ * C.transpose()) * S_inv * (C * P_);  
 }
-void KalmanFilter::updateEKF(const Eigen::MatrixXf &y,
-                             const Eigen::MatrixXf &y_pred,
-                             const Eigen::MatrixXf &C)
+void KalmanFilter::updateEKF(const Eigen::MatrixXd &y,
+                             const Eigen::MatrixXd &y_pred,
+                             const Eigen::MatrixXd &C)
 {
   updateEKF(y, y_pred, C, R_);
 }
-void KalmanFilter::update(const Eigen::MatrixXf &y, const Eigen::MatrixXf &C,
-                          const Eigen::MatrixXf &R)
+void KalmanFilter::update(const Eigen::MatrixXd &y, const Eigen::MatrixXd &C,
+                          const Eigen::MatrixXd &R)
 {
-  const Eigen::MatrixXf y_pred = C * x_;
+  const Eigen::MatrixXd y_pred = C * x_;
   updateEKF(y, y_pred, C, R);
 }
-void KalmanFilter::update(const Eigen::MatrixXf &y) { update(y, C_, R_); }
+void KalmanFilter::update(const Eigen::MatrixXd &y) { update(y, C_, R_); }
